Move session file functions to `session` from `utils`
